refactor: menu, menugroup 패키지 병합
