+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -1,7 +1,6 @@
 package kitchenpos.domain.menu;
 
 import kitchenpos.common.BaseDate;
-import kitchenpos.domain.menugroup.MenuGroup;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
