+++ src/main/java/kitchenpos/domain/menu/MenuGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.domain.menugroup;
+package kitchenpos.domain.menu;
 
-import kitchenpos.domain.menugroup.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
