+++ src/test/java/kitchenpos/domain/MenuValidatorTest.java
@@ -2,7 +2,7 @@ package kitchenpos.domain;
 
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuValidator;
-import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
